reorganize product shape
