--- src/page-components/folder/query.js
@@ -1,16 +1,0 @@
-import fragments from 'lib/graph/fragments';
-
-export default `
-  query FOLDER_PAGE($language: String!, $path: String, $version: VersionLabel!) {
-    folder: catalogue(language: $language, path: $path, version: $version) {
-      ...item
-
-      children {
-        ...item
-        ...product
-      }
-    }
-  }
-
-  ${fragments}
-`;
