+++ src/shapes/product/components/variant-selector/styles.js
